api/requirements.txt
absl-py==1.4.0
accelerate==0.21.0 
addict==2.4.0
annotated-types==0.5.0
anyio==3.7.1
audioread==3.0.0
av==10.0.0
basicsr==1.4.2
cachetools==5.3.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
cmake==3.27.2
contourpy==1.1.0
cycler==0.11.0
decorator==5.1.1
diffusers==0.20.0
exceptiongroup==1.1.3
face-alignment==1.3.5
facexlib==0.3.0
fastapi==0.101.1
ffmpeg-python==0.2.0
ffprobe-python==1.0.3
filelock==3.12.2
filterpy==1.4.5
fonttools==4.42.1
fsspec==2023.6.0
future==0.18.3
gfpgan==1.3.8
h11==0.14.0
huggingface-hub==0.16.4
idna==3.4
imageio==2.19.3
imageio-ffmpeg==0.4.7
importlib-metadata==6.8.0
importlib-resources==6.0.1
Jinja2==3.1.2
joblib==1.1.0
kiwisolver==1.4.4
kornia==0.6.8
librosa==0.9.2
lit==16.0.6
llvmlite==0.40.1
lmdb==1.4.1
Markdown==3.4.4
MarkupSafe==2.1.3
matplotlib==3.7.2
mpmath==1.3.0
networkx==3.1
numba==0.57.1
numpy==1.23.4
opencv-python==4.8.0.76
packaging==23.1
Pillow==10.0.0
platformdirs==3.10.0
pooch==1.7.0
protobuf==4.23.4
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.2.1
pydantic_core==2.6.1
pydub==0.25.1
pyparsing==3.0.9
python-dateutil==2.8.2
python-multipart==0.0.6
PyWavelets==1.4.1
PyYAML==6.0.1
regex==2023.8.8
requests==2.31.0
requests-oauthlib==1.3.1
resampy==0.3.1
rsa==4.9
safetensors==0.3.2
scikit-image==0.19.3
scikit-learn==1.1.3
scipy==1.10.1
six==1.16.0
sniffio==1.3.0
soundfile==0.12.1
starlette==0.27.0
sympy==1.12
tb-nightly==2.15.0a20230822
tensorboard-data-server==0.7.1
threadpoolctl==3.2.0
tifffile==2023.8.12
tokenizers==0.13.3
tomli==2.0.1
tqdm==4.66.1
transformers==4.32.0
typing_extensions==4.7.1
urllib3==1.26.16
uvicorn==0.23.2
Werkzeug==2.3.7
yacs==0.1.8
yapf==0.40.1
zipp==3.16.2